+++ Sources/Flare/Classes/Models/Internal/SK1StoreProduct.swift
@@ -42,6 +42,10 @@ extension SK1StoreProduct: ISKProduct {
     var currencySymbol: String? {
         numberFormatter.currencySymbol
     }
+    
+    var locale: Locale {
+        product.priceLocale
+    }
 
     var price: Decimal {
         product.price as Decimal