+++ Sources/Flare/Classes/Models/Internal/SK2StoreProduct.swift
@@ -49,6 +49,10 @@ extension SK2StoreProduct: ISKProduct {
     var currencySymbol: String? {
         numberFormatter.currencySymbol
     }
+    
+    var locale: Locale {
+        currencyFormat.locale
+    }
 
     var price: Decimal {
         product.price