+++ Sources/Flare/Classes/Models/StoreProduct.swift
@@ -65,6 +65,10 @@ extension StoreProduct: ISKProduct {
     public var currencySymbol: String? {
         product.currencySymbol
     }
+    
+    public var locale: Locale {
+        product.locale
+    }
 
     public var price: Decimal {
         product.price